Muita coisa

+++ Entities/FuncionarioRespostasQuestionarioSf36.cs
@@ -2,46 +2,47 @@
 {
     public class FuncionarioRespostasQuestionarioSf36
     {
-        public int Question1 { get; set; }
+        public double Question1 { get; set; }
+
         public string Question2 { get; set; }
-        public int Question3 { get; set; }
-        public int Question3a { get; set; }
-        public int Question3b { get; set; }
-        public int Question3c { get; set; }
-        public int Question3d { get; set; }
-        public int Question3e { get; set; }
-        public int Question3f { get; set; }
-        public int Question3g { get; set; }
-        public int Question3h { get; set; }
-        public int Question3i { get; set; }
-        public int Question3j { get; set; }
-        public int Question4 { get; set; }
-        public int Question4a { get; set; }
-        public int Question4b { get; set; }
-        public int Question4c { get; set; }
-        public int Question4d { get; set; }
-        public int Question5 { get; set; }
-        public int Question5a { get; set; }
-        public int Question5b { get; set; }
-        public int Question5c { get; set; }
-        public int Question6 { get; set; }
-        public int Question7 { get; set; }
-        public int Question8 { get; set; }
-        public int Question9 { get; set; }
-        public int Question9a { get; set; }
-        public int Question9b { get; set; }
-        public int Question9c { get; set; }
-        public int Question9d { get; set; }
-        public int Question9e { get; set; }
-        public int Question9f { get; set; }
-        public int Question9g { get; set; }
-        public int Question9h { get; set; }
-        public int Question9i { get; set; }
-        public int Question10 { get; set; }
-        public int Question11 { get; set; }
-        public int Question11a { get; set; }
-        public int Question11b { get; set; }
-        public int Question11c { get; set; }
-        public int Question11d { get; set; }
+        public double Question3 { get; set; }
+        public double Question3a { get; set; }
+        public double Question3b { get; set; }
+        public double Question3c { get; set; }
+        public double Question3d { get; set; }
+        public double Question3e { get; set; }
+        public double Question3f { get; set; }
+        public double Question3g { get; set; }
+        public double Question3h { get; set; }
+        public double Question3i { get; set; }
+        public double Question3j { get; set; }
+        public double Question4 { get; set; }
+        public double Question4a { get; set; }
+        public double Question4b { get; set; }
+        public double Question4c { get; set; }
+        public double Question4d { get; set; }
+        public double Question5 { get; set; }
+        public double Question5a { get; set; }
+        public double Question5b { get; set; }
+        public double Question5c { get; set; }
+        public double Question6 { get; set; }
+        public double Question7 { get; set; }
+        public double Question8 { get; set; }
+        public double Question9 { get; set; }
+        public double Question9a { get; set; }
+        public double Question9b { get; set; }
+        public double Question9c { get; set; }
+        public double Question9d { get; set; }
+        public double Question9e { get; set; }
+        public double Question9f { get; set; }
+        public double Question9g { get; set; }
+        public double Question9h { get; set; }
+        public double Question9i { get; set; }
+        public double Question10 { get; set; }
+        public double Question11 { get; set; }
+        public double Question11a { get; set; }
+        public double Question11b { get; set; }
+        public double Question11c { get; set; }
+        public double Question11d { get; set; }
     }
 }
